+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add comprehensive test coverage
 * Fix the bug for `AddShipmentsToBatch`, `AddShipmentsToBatchWithContext`, change the parameter from `string` to `interface{}` so user can pass in either Shipment IDs or Shipment structs
 * Fix the bug for `RemoveShipmentsFromBatch`, and `RemoveShipmentsFromBatchWithContext`, change the parameter from `*Shipment` to `interface{}` so user can pass in either Shipment IDs or Shipment structs
+* Add `CreateRefund`, `CreateRefundWithContext`, `ListRefunds`, `ListRefundsWithContext`, `GetRefund`, and `GetRefundWithContext` in the Refund file
 
 ## v2.0.1 (2022-02-10)